backend/requirements.txt
asgiref==3.4.1
Django==3.2.8
django-cleanup==6.0.0
django-extensions==3.1.5
djangorestframework==3.13.1
Pillow==9.0.1
pytz==2021.3
sqlparse==0.4.2
